lab_software/dinamica_xp/requirements.txt
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
IPython==8.2.0
rich==13.7.1

supabase==2.4.1
python-environ==0.4.54
psycopg2==2.9.9